@@ -16,30 +16,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "v6-lts",
-    "@angular/common": "v6-lts",
-    "@angular/compiler": "v6-lts",
-    "@angular/core": "v6-lts",
-    "@angular/forms": "v6-lts",
-    "@angular/platform-browser": "v6-lts",
-    "@angular/platform-browser-dynamic": "v6-lts",
+    "@angular/animations": "8.2.14",
+    "@angular/common": "8.2.14",
+    "@angular/compiler": "8.2.14",
+    "@angular/core": "8.2.14",
+    "@angular/forms": "8.2.14",
+    "@angular/platform-browser": "8.2.14",
+    "@angular/platform-browser-dynamic": "8.2.14",
     "ajv": "^6.9.2",
     "cometd": "^3.1.3",
     "core-js": "^2.5.3",
-    "rxjs": "^6.5.5",
+    "rxjs": "^6.6.7",
+    "tslib": "^1.10.0",
     "url-polyfill": "^1.1.8",
     "url-search-params-polyfill": "^8.0.0",
     "web-animations-js": "^2.3.2",
-    "zone.js": "^0.10.3"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "v6-lts",
-    "@angular/cli": "v6-lts",
-    "@angular/compiler-cli": "v6-lts",
+    "@angular-devkit/build-angular": "~0.803.29",
+    "@angular/cli": "8.3.29",
+    "@angular/compiler-cli": "8.2.14",
     "@types/node": "^9.4.6",
-    "codelyzer": "latest",
+    "codelyzer": "^5.0.1",
     "font-awesome": "^4.7.0",
     "tslint": "latest",
-    "typescript": "2.9.2"
+    "typescript": "3.5.3"
   }
 }
